ottawa_chatbot/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
httpx==0.27.2
openai==1.40.6
pypdf==4.3.1

